Implement enemy melee attack logic and player damage
Added attack logic to EnergyMeleeAI, allowing enemies to attack the player when in range and applying cooldowns. Introduced TakeDamage method in PlayerStats to handle health reduction and death. Changed Enemy.Attack.attackDamage type from float to int for consistency.

+++ Assets/Scripts/ScriptableObjectScripts/Enemy.cs
@@ -10,7 +10,7 @@ public class Enemy : ScriptableObject
     public class Attack
     {
         public AttackType attackType = AttackType.Melee;
-        public float attackDamage = 20f;
+        public int attackDamage = 20;
         
         [Tooltip("How close the enemy needs to be to attack")]
         public float attackRange = 10f;
